improved ds
