+++ infrastructure/package.json
@@ -44,6 +44,7 @@
   },
   "dependencies": {
     "@aws-cdk/aws-lambda-python-alpha": "2.132.0-alpha.0",
+    "@aws-sdk/client-cloudwatch": "^3.572.0",
     "@aws-sdk/client-dynamodb": "^3.569.0",
     "@aws-sdk/client-iam": "^3.569.0",
     "@aws-sdk/client-kms": "^3.569.0",
