+++ infrastructure/src/main.ts
@@ -25,7 +25,6 @@ import { RestApiGateway } from './shared/rest-api-gateway';
 import { SnsTopics } from './shared/sns-topics';
 import { TrainDeploy } from './shared/train-deploy';
 import { ESD_VERSION } from './shared/version';
-import { DashboardStack } from "./shared/dashboard";
 
 const app = new App();
 
@@ -115,8 +114,6 @@ export class Middleware extends Stack {
     );
 
     const ddbTables = new Database(this, 'sd-ddb');
-
-    new DashboardStack(this);
 
     const commonLayers = new LambdaCommonLayer(this, 'sd-common-layer');
